Chore: rename Expyrementor to XpyriMentor
