--- Expyrementor/__init__.py
@@ -1,3 +1,0 @@
-from .expyrementor import Expyrementor
-
-__all__ = ["Expyrementor"]
