+++ ProcessOptimizer/tests/test_suggestors/test_initial_point_suggestor.py
@@ -1,5 +1,5 @@
 import numpy as np
-from Expyrementor.suggestors import (
+from XpyriMentor.suggestors import (
     suggestor_factory,
     InitialPointStrategizer,
     DefaultSuggestor,
