+++ XpyriMentor/__init__.py
@@ -1,0 +1,3 @@
+from .xpyrimentor import XpyriMentor
+
+__all__ = ["XpyriMentor"]
